For Loop trouble
Download function seems to continue opening pdf chart pages before the URL of the pdf is retrieved on each page; the for loop on line 49 continues to loop before the nested for loop on line 58 is finished. The driver can only work on one page at a time, so I need to stop the function from opening another page until the 'View Full PDF' URL is retrieved from the active tab.