Working Code, but..
I also attached the text file that is generated by calling that function. This is an embarrassing question, but I cannot figure out where the PDF file to read is even being passed in the function. It only reads in one of the PDFs from that bucket, but I never specified which one.