Full Scope
All the code I wrote for the project with comments/docstrings along the way. Happy to clarify anything. The second section to read the PDF data and upload it to SQL should be fully functional other than reconnecting to a different database, but the first section to download the PDFs from Equibase will have to be mostly reworked since that package is not functional.